# Improve frame_sync_impl::get_symbol_val() performance