On Apple M3 hyperfine shows 22% of improvement:
```
Benchmark 1: python3 gr_meshtastic.py
  Time (mean ± σ):     428.8 ms ±   6.7 ms    [User: 420.3 ms, System: 81.5 ms]
  Range (min … max):   420.8 ms … 439.1 ms    10 runs
```
```
Benchmark 1: python3 gr_meshtastic.py
  Time (mean ± σ):     335.1 ms ±   1.6 ms    [User: 326.2 ms, System: 79.8 ms]
  Range (min … max):   332.7 ms … 337.0 ms    10 runs
```

On RPi4 the improvement is about 13% (8.12s to 7.09s). Some details on the benchmarking setup are in https://github.com/tapparelj/gr-lora_sdr/issues/92.

There are other places in the code where a state returned by kiss_fft_alloc() is not reused, but it seems that they are not on the hot path.