@mskvortsov 2024-05-03
Also fixed minor imperfections found by valgrind.